# backend/src/docs/swagger/index.yaml
openapi: 3.0.0
info:
  title: Social Developer API
  version: 1.0.0
  description: |
    API documentation for the Social Developer platform.
    This API provides endpoints for managing social media automation tasks, payments, and user accounts.
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.socialdeveloper.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error, fail]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        balance:
          type: number
          format: float
        userMode:
          type: string
          enum: [taskDoer, taskGiver]
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...